Fixed bug in postgres implementation regarding saving longs. Added support for LocalDate, ZonedDateTime and LocalTime. Updated readme and added newer tests. closes #2

+++ src/main/resources/sql/postgresql/createStoreLong.sql
@@ -14,7 +14,7 @@
 CREATE TABLE JdsStoreLong(
 	FieldId     BIGINT,
 	EntityGuid  VARCHAR(48),
-	Value       INTEGER,
+	Value       BIGINT,
 	PRIMARY KEY (FieldId,EntityGuid),
 	FOREIGN KEY (EntityGuid) REFERENCES JdsStoreEntityOverview(EntityGuid) ON DELETE CASCADE
 );
